refactor(master): add apply drivers add run cmd and vendor (#902)
* refactor(master): fix var named

* refactor(master): add apply drivers

* refactor(master): add apply drivers add run cmd and vendor

* refactor(master): add apply drivers add run cmd and vendor

* refactor(master): add apply drivers add run cmd and vendor

+++ pkg/types/v1beta1/cluster.go
@@ -36,7 +36,7 @@ type ClusterSSH struct {
 	Passwd   string `json:"passwd,omitempty"`
 	Pk       string `json:"pk,omitempty"`
 	PkPasswd string `json:"pkPasswd,omitempty"`
-	Port     int32  `json:"port,omitempty"`
+	Port     uint16 `json:"port,omitempty"`
 }
 
 type ClusterHost struct {
